move DBLoader to forest.drivers.unified_model.Loader
